Add simple authorization checking

+++ app/repositories/appointments.rb
@@ -10,31 +10,44 @@ class Appointments
       Appointment.find(id)
     end
 
-    def find_by_user(user_id)
-      Appointment.where(user_id:)
+    def find_by_user(user)
+      Appointment.where(user_id: user.id)
     end
 
-    def create(params)
-      # validate params
-      params = params.to_h.symbolize_keys
-      params => {user_id: user_id, title: title, desc: description, start: start_date, end: end_date}
+    def create(user, params)
+      create_params = populate_attributes(params).merge(user_id: user.id)
 
-      Appointment.create!(title:, description:, start_date:, end_date:, user_id:)
+      Appointment.create!(create_params)
     end
 
-    def update(id, params)
-      # validate params
+    def update(id, user, params)
       appointment = find(id)
 
-      params = params.to_h.symbolize_keys
-      params => {user_id: user_id, title: title, desc: description, start: start_date, end: end_date}
+      raise 'Access Denied' if user.id != appointment.user_id
 
-      appointment.update!(title:, description:, start_date:, end_date:, user_id:)
+      appointment.update!(populate_attributes(params))
     end
 
-    def destroy(id)
+    def destroy(id, user)
       appointment = find(id)
+
+      raise 'Access Denied' if user.id != appointment.user_id
+
       appointment.destroy!
+    end
+
+    private
+
+    def populate_attributes(params)
+      params.to_h.symbolize_keys => {
+        user_id: user_id,
+        title: title,
+        desc: description,
+        start: start_date,
+        end: end_date
+      }
+
+      { user_id:, title:, description:, start_date:, end_date: }
     end
   end
 end
